requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-JWT-Extended
Flask-Mail
python-dotenv
requests
gunicorn
email-validator
bcrypt
